requirements.txt
pandas
numpy
matplotlib
jupyter
tensorflow_datasets
flask
pytest
sklearn